Added more functions
