+++ webAO/dom/iniEdit.ts
@@ -1,0 +1,14 @@
+import { client } from "../client";
+import { packetHandler } from "../packets/packetHandler";
+
+/**
+ * Triggered by the ini button.
+ */
+export async function iniedit() {
+    const ininame = (<HTMLInputElement>document.getElementById("client_ininame"))
+        .value;
+    const inicharID = client.charID;
+    await client.handleCharacterInfo(ininame.split("&"), inicharID);
+    packetHandler.get("PV")!(`PV#0#CID#${inicharID}`.split("#"));
+}
+window.iniedit = iniedit;
