+++ webAO/dom/modCallTest.ts
@@ -1,0 +1,8 @@
+import { packetHandler } from "../packets/packetHandler";
+/**
+ * Triggered by the modcall sfx dropdown
+ */
+export function modcall_test() {
+    packetHandler.get("ZZ")!("test#test".split("#"));
+}
+window.modcall_test = modcall_test;
