+++ webAO/dom/toggleShout.ts
@@ -1,6 +1,5 @@
 import { selectedShout, setSelectedShout } from "../client";
 
-
 /**
  * Highlights and selects a shout for in-character chat.
  * If the same shout button is selected, then the shout is canceled.
@@ -9,14 +8,14 @@ import { selectedShout, setSelectedShout } from "../client";
 export function toggleShout(shout: number) {
     if (shout === selectedShout) {
         document.getElementById(`button_${shout}`)!.className = "client_button";
-        selectedShout = 0;
+        setSelectedShout(0);
     } else {
         document.getElementById(`button_${shout}`)!.className = "client_button dark";
         if (selectedShout) {
             document.getElementById(`button_${selectedShout}`)!.className =
                 "client_button";
         }
-        selectedShout = shout;
+        setSelectedShout(shout);
     }
 }
 window.toggleShout = toggleShout;
